cleanup account-view
